wip: improve in mem structure.
